Fix out-of-source builds
When building something out-of-source, make's VPATH mechanism is used so that
files in the source directory are also found. This makes make automatically
check that other directory for needed files and use those when needed.

However, make doesn't magically copy those files over. We have to explicitly use
the file that make found. Do so by using $< for refering to shape_to_id.gperf
instead of hardcoding the name of the source file.

This fixes out-of-source git builds.

Out-of-source release builds already worked before this, because releases come
with a pre-generated shape_to_id.c.

This fixes https://bugs.freedesktop.org/show_bug.cgi?id=80153

Signed-off-by: Uli Schlachter <psychon@znc.in>