Perform safety check before trying to load glyph cursor
The passed in cursor name to xcb_cursor_load_cursor might not match
one of the predefined font cursor values. Without the check the
call to create glyph cursor will fail with a BadValue error, but
the library returns the id allocated for the xcb_cursor_t. A user
of the library gets a value which looks like a valid cursor, but
when using it for e.g. a cursor value in xcb_create_window it raises
a BadCursor error.