documentation: Call xcb_free_cursor() when done
The documentation for xcb_cursor_load_cursor() say that the resulting
cursor must be freed, but the example in the documentation does not do
so. It does not matter much for this example, but I think it is a good
idea to use it to hint to this requirement. Also, the example already
uses xcb_cursor_context_free(), so it tries to clean up after itself.

Signed-off-by: Uli Schlachter <psychon@znc.in>